pyproject.toml
[project]
name = "codesage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.24",
    "langchain-core>=0.3.55",
    "langchain-google-vertexai>=2.0.20",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.3.31",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
